requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
audioread==3.0.1
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
colorclass==2.2.2
coloredlogs==15.0.1
compressed_rtf==1.0.6
contourpy==1.3.1
cryptography==44.0.0
curl_cffi==0.6.3
cycler==0.12.1
dataclasses-json==0.5.14
decorator==5.1.1
Deprecated==1.2.15
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
duckduckgo_search==6.3.7
durationpy==0.9
easygui==0.98.3
ebcdic==1.1.1
einops==0.8.0
emoji==2.14.0
encodec==0.1.1
eventlet==0.38.0
extract-msg==0.45.0
faiss-cpu==1.9.0.post1
fastapi==0.115.5
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-SocketIO==5.4.1
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.4.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
googleapis-common-protos==1.66.0
gpt4all==1.0.8
greenlet==3.0.3
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.7
IMAPClient==2.3.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
jsonargparse==4.34.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.0.274
langchain-community==0.3.8
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langsmith==0.0.92
lark==1.1.9
lazy_loader==0.4
librosa==0.10.2.post1
lightning-utilities==0.11.9
llama_cpp_python==0.3.2
llvmlite==0.43.0
loguru==0.7.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.9.2
mdurl==0.1.2
mecab-python3==1.0.10
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
msoffcrypto-tool==5.4.2
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.14.2
natsort==8.4.0
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numexpr==2.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.46
oletools==0.60.2
ollama==0.1.9
onnxruntime==1.20.1
openai==1.23.6
openai-whisper==20240930
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.1
outetts==0.2.1
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandoc==2.3
pcodedmp==1.2.6
pesq==0.0.4
pillow==10.3.0
platformdirs==4.3.6
playwright==1.43.0
plumbum==1.9.0
ply==3.11
polars==1.16.0
pooch==1.8.2
posthog==3.7.4
primp==0.8.1
protobuf==4.25.5
pyarrow==16.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.0b1
pyee==11.1.0
Pygments==2.17.2
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pypandoc==1.11
pyparsing==3.2.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-engineio==4.10.1
python-magic==0.4.27
python-socketio==5.11.4
pytorch-lightning==2.4.0
pytz==2024.1
PyYAML==6.0.1
red-black-tree-mod==1.20
referencing==0.35.0
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
resampy==0.4.3
RestrictedPython==7.4
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
RTFDE==0.1.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
setuptools==75.6.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.5.0.post1
SQLAlchemy==2.0.29
starlette==0.41.3
streamlit==1.40.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
toolz==0.12.1
torch==2.5.1
torchaudio==2.5.1
torchcrepe==0.0.23
torchmetrics==1.6.0
torchvision==0.20.1
tornado==6.4
tqdm==4.66.1
transformers==4.46.3
triton==3.1.0
typeguard==4.4.1
typer==0.13.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
unidic-lite==1.0.8
unstructured==0.10.8
urllib3==2.0.4
uroman==1.3.1.1
uvicorn==0.32.1
uvloop==0.21.0
watchdog==5.0.3
watchfiles==1.0.0
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.21.0
